Under GPL v3 license, we herein open source of Music Go for all music lovers (not including the back-end service hosted by other parties). Hope you enjoy music too. ^_^

+++ app/src/main/java/code/name/monkey/retromusic/volume/OnAudioVolumeChangedListener.kt
@@ -14,5 +14,5 @@
 package code.name.monkey.retromusic.volume
 
 interface OnAudioVolumeChangedListener {
-    fun onAudioVolumeChanged(currentVolume: Int, maxVolume: Int)
+    fun onAudioVolumeChanged(currentVolume: Float, maxVolume: Float)
 }
